lib/infra/pipe: use 'int' for size and pend
Reduce obejct size + cleanups.

+++ include/silofs/infra/pipe.h
@@ -23,8 +23,8 @@ struct iovec;
 
 struct silofs_pipe {
 	int fd[2];
-	size_t size;
-	size_t pend;
+	int size;
+	int pend;
 };
 
 struct silofs_nilfd {
